+++ rlkit/torch/model_based/dreamer/experiments/raps_experiment.py
@@ -9,7 +9,10 @@ def experiment(variant):
 
     import rlkit.envs.primitives_make_env as primitives_make_env
     import rlkit.torch.pytorch_util as ptu
-    from rlkit.envs.mujoco_vec_wrappers import DummyVecEnv, StableBaselinesVecEnv
+    from rlkit.envs.wrappers.mujoco_vec_wrappers import (
+        DummyVecEnv,
+        StableBaselinesVecEnv,
+    )
     from rlkit.torch.model_based.dreamer.actor_models import (
         ActorModel,
         ConditionalActorModel,
