+++ rlkit/torch/model_based/plan2explore/experiments/dmc_plan2explore_experiment.py
@@ -138,7 +138,7 @@ def experiment(variant):
     import torch
 
     import rlkit.torch.pytorch_util as ptu
-    from rlkit.envs.mujoco_vec_wrappers import DummyVecEnv
+    from rlkit.envs.wrappers.mujoco_vec_wrappers import DummyVecEnv
     from rlkit.torch.model_based.dreamer.actor_models import ActorModel
     from rlkit.torch.model_based.dreamer.dreamer_policy import (
         ActionSpaceSamplePolicy,
